+++ edison-core/src/main/java/de/otto/edison/status/domain/ServiceDependency.java
@@ -1,14 +1,11 @@
 package de.otto.edison.status.domain;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
 import de.otto.edison.annotations.Beta;
 import net.jcip.annotations.Immutable;
 
 import java.util.List;
 import java.util.Objects;
 
-import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 import static java.util.Collections.emptyList;
 import static java.util.Objects.hash;
 import static java.util.Objects.requireNonNull;
@@ -20,8 +17,6 @@ import static java.util.Objects.requireNonNull;
  */
 @Beta
 @Immutable
-@JsonIgnoreProperties(ignoreUnknown = true)
-@JsonInclude(NON_NULL)
 public class ServiceDependency extends ExternalDependency {
     public static final String TYPE_SERVICE = "service";