+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * **[edison-core]** Deprecated the (Beta) ServiceSpecs.
 
 **New Features:**
+* **[edison-core]** Added a UI to configure the log levels of the service. The UI is added to the right nav bar and is
+available under `/internal/loggers`
 * **[edison-core]** Introduced the configuration of external dependencies, which is replacing the now
 deprecated ServiceSpecs. External dependencies include datasources like databases or queues, as well as
 dependencies to REST or other services.
@@ -19,8 +21,6 @@ dependencies to REST or other services.
 * **[edison-mongo]** Fixed display of mongo password in /internal/env
 
 **New Features:**
-* **[edison-core]** Added a UI to configure the log levels of the service. The UI is added to the right nav bar and is
-available under `/internal/loggers`
 * **[edison-core]** Allow LDAP authentication for user-defined paths.
 * **[edison-mongo]** Add @ConditionalOnMissingBean to MongoTogglzRepository
 * **[edison-mongo]** Make ID and ETAG constants public to be able to access them from outside